Begin extracting `PromptLibrary`
