+++ crates/assistant/src/assistant_panel.rs
@@ -1,6 +1,5 @@
 use crate::{
-    humanize_token_count, prompt_library::open_prompt_library,
-    slash_command::SlashCommandCompletionProvider, slash_command_picker,
+    humanize_token_count, slash_command::SlashCommandCompletionProvider, slash_command_picker,
     terminal_inline_assistant::TerminalInlineAssistant, Assist, AssistantPatch,
     AssistantPatchStatus, CacheStatus, ConfirmCommand, Content, Context, ContextEvent, ContextId,
     ContextStore, ContextStoreEvent, CopyCode, CycleMessageRole, DeployHistory,
@@ -55,7 +54,7 @@ use multi_buffer::MultiBufferRow;
 use picker::{Picker, PickerDelegate};
 use project::lsp_store::LocalLspAdapterDelegate;
 use project::{Project, Worktree};
-use prompt_library::PromptBuilder;
+use prompt_library::{open_prompt_library, PromptBuilder};
 use rope::Point;
 use search::{buffer_search::DivRegistrar, BufferSearchBar};
 use serde::{Deserialize, Serialize};
@@ -1186,7 +1185,18 @@ impl AssistantPanel {
     }
 
     fn deploy_prompt_library(&mut self, _: &DeployPromptLibrary, cx: &mut ViewContext<Self>) {
-        open_prompt_library(self.languages.clone(), cx).detach_and_log_err(cx);
+        open_prompt_library(
+            self.languages.clone(),
+            Arc::new(|| {
+                Box::new(SlashCommandCompletionProvider::new(
+                    Arc::new(SlashCommandWorkingSet::default()),
+                    None,
+                    None,
+                ))
+            }),
+            cx,
+        )
+        .detach_and_log_err(cx);
     }
 
     fn toggle_model_selector(&mut self, _: &ToggleModelSelector, cx: &mut ViewContext<Self>) {
