+++ crates/prompt_library/Cargo.toml
@@ -14,8 +14,10 @@ path = "src/prompt_library.rs"
 [dependencies]
 anyhow.workspace = true
 assets.workspace = true
+assistant_slash_command.workspace = true
 chrono.workspace = true
 collections.workspace = true
+editor.workspace = true
 fs.workspace = true
 futures.workspace = true
 fuzzy.workspace = true
@@ -23,11 +25,20 @@ gpui.workspace = true
 handlebars.workspace = true
 heed.workspace = true
 language.workspace = true
+language_model.workspace = true
 log.workspace = true
+menu.workspace = true
 parking_lot.workspace = true
 paths.workspace = true
+picker.workspace = true
+release_channel.workspace = true
 rope.workspace = true
 serde.workspace = true
+settings.workspace = true
 text.workspace = true
+theme.workspace = true
+ui.workspace = true
 util.workspace = true
 uuid.workspace = true
+workspace.workspace = true
+zed_actions.workspace = true
